Add files via upload
K-Nearest Neighbor (KNN) is a classification model. This attempt to estimate the conditional distribution of Y given X, and then classify a given observation to the class with highest estimated probability. Given a positive in- K-nearest integer K and a test observation x0, the KNN classifier first identifies the neighbors K points in the training data that are closest to x0,

Reference : https://www.ime.unicamp.br/~dias/Intoduction%20to%20Statistical%20Learning.pdf